// types.ts
export interface WeatherData {
    temperature: number;
    feels_like: number;
    wind_speed: number;
    wind_direction: number;
    cloud_cover: number;
    icon: string;
    description: string;
    forecast: { time: string; temperature: number }[];
}